refactor(mcp): extract catalog operations
